chatbot_backend/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-mistralai>=0.2.0
faiss-cpu>=1.9.0
duckduckgo-search>=6.1.7
python-dotenv